requirements.txt
requests
feedparser
beautifulsoup4
python-telegram-bot==13.15
langdetect
